requirements.txt
fpdf2==2.8.2
lxml==4.9.1
matplotlib==3.8.2
numpy==1.24.2
pandas==2.0.0
scikit_learn==1.4.1.post1
spacy==3.7.4
spacy_syllables==3.0.2
